Change monorepo structure to normal repo structure
