--- packages/preprocessor/package.json
@@ -1,28 +1,0 @@
-{
-	"name": "@brefer/preprocessor",
-	"version": "2.0.0",
-	"description": "A preprocessor for Svelte that allows you to use a shortened syntax for reactivity.",
-	"main": "./src/index.js",
-	"type": "module",
-	"types": "./types/index.d.ts",
-	"files": [
-		"src",
-		"types"
-	],
-	"exports": {
-		".": {
-			"types": "./types/index.d.ts",
-			"default": "./src/index.js"
-		}
-	},
-	"scripts": {
-		"generate:types": "dts-buddy -m \"@brefer/preprocessor:src/public.d.ts\""
-	},
-	"keywords": [],
-	"author": "",
-	"license": "ISC",
-	"dependencies": {
-		"@babel/parser": "^7.23.9",
-		"recast": "^0.23.4"
-	}
-}
