--- packages/preprocessor/src/public.d.ts
@@ -1,2 +1,0 @@
-export * from "./index.js";
-import "./ambient.js";
