--- packages/preprocessor/tsconfig.json
@@ -1,4 +1,0 @@
-{
-	"include": ["src"],
-	"extends": "../../tsconfig.base.json"
-}
