--- packages/vite-plugin-svelte/package.json
@@ -1,22 +1,0 @@
-{
-	"name": "@brefer/vite-plugin-svelte",
-	"private": true,
-	"version": "1.0.0",
-	"type": "module",
-	"types": "./types/index.d.ts",
-	"exports": {
-		".": {
-			"types": "./types/index.d.ts",
-			"default": "./src/index.js"
-		}
-	},
-	"scripts": {
-		"generate:types": "dts-buddy -m \"@brefer/vite-plugin-svelte:src/public.d.ts\""
-	},
-	"dependencies": {
-		"@brefer/preprocessor": "workspace:^"
-	},
-	"devDependencies": {
-		"vite": "^5.1.1"
-	}
-}
