--- packages/vite-plugin-svelte/tsconfig.json
@@ -1,4 +1,0 @@
-{
-	"include": ["src"],
-	"extends": "../../tsconfig.base.json"
-}
