+++ src/ambient.d.ts
@@ -1,21 +1,5 @@
-declare module '@brefer/preprocessor' {
-	/**
-	 * Preprocesses the content of the script tag in .svelte files.
-	 * @param content - The content of the script tag
-	 * @param filename - The name of the file
-	 * @param lang - The name of the file
-	 */
-	export function preprocessScript(content: string, filename?: string | undefined, lang?: string | boolean | undefined): import("recast/lib/printer.js").PrintResultType;
-	/**
-	 * Preprocessor for Brefer syntax, using variable prefixes to handle reactivity.
-	 * It avoids the need to call `$state`, `$derived` or `$effect` runes every time.
-	 *
-	 * If you also want to preprocess .svelte.js files, use `@brefer/vite-plugin-svelte` instead.
-	 *
-	 * */
-	export function breferPreprocess(): import("svelte/compiler").PreprocessorGroup;
-}/**
- * Corresponds to Svelte 5's `$derived` rune.
+/**
+ * Corresponds to Svelte 5's `$derived` rune. If a function is passed as an argument, `$derived.call` will be used instead.
  *
  * Declares derived state, i.e. one that depends on other state variables.
  * The expression inside `$(...)` should be free of side-effects.
@@ -30,25 +14,6 @@ declare module '@brefer/preprocessor' {
  * @param expression The derived state expression
  */
 declare function $<T>(expression: T): T;
-/**
- * Corresponds to Svelte 5's `$derived.call` rune.
- *
- * Sometimes you need to create complex derivations that don't fit inside a short expression.
- * In these cases, you can use `$derived.call` which accepts a function as its argument.
- *
- * Example:
- * ```ts
- * let total = $derived.call(() => {
- *   let result = 0;
- *	 for (const n of numbers) {
- *	   result += n;
- *   }
- *   return result;
- * });
- * ```
- *
- * https://svelte-5-preview.vercel.app/docs/runes#$derived-call
- */
 declare function $<T>(fn: () => T): T;
 
 /**
@@ -188,5 +153,3 @@ declare function $static<T>(value: T): T;
  */
 declare function $frozen<T>(initial: T): Readonly<T>;
 declare function $frozen<T>(): Readonly<T> | undefined;
-
-//# sourceMappingURL=index.d.ts.map
