+++ src/handlers/declaration.js
