+++ src/index.js
@@ -1,11 +1,29 @@
 import { preprocess as sveltePreprocess } from "svelte/compiler";
-import { breferPreprocess, preprocessScript } from "@brefer/preprocessor";
+import { preprocessScript } from "./preprocess.js";
 import { createFilter } from "vite";
+
+/**
+ * Preprocessor for Brefer syntax, using variable prefixes to handle reactivity.
+ * It avoids the need to call `$state`, `$derived` or `$effect` runes every time.
+ *
+ * If you also want to preprocess .svelte.js files, use `brefer` instead.
+ *
+ * @returns { import("svelte/compiler").PreprocessorGroup }
+ */
+export function breferPreprocess() {
+	return {
+		name: "brefer-preprocessor",
+		async script({ content, filename, attributes }) {
+			return preprocessScript(content, filename, attributes.lang);
+		},
+	};
+}
+
 
 /**
  * Brefer vite plugin for svelte. It allows to preprocess .svelte.js files as well as .svelte files.
  *
- * Prefer the use of `@brefer/preprocessor` if you want to preprocess .svelte files only.
+ * Prefer the use of `breferPreprocess` if you want to preprocess .svelte files only.
  *
  * @export
  * @param {import("./public.js").BreferConfig} config
@@ -33,7 +51,7 @@ export function brefer(config = {}) {
 				};
 			}
 			if (id.endsWith(".svelte.js") || id.endsWith(".svelte.ts")) {
-				const preprocessed = await preprocessScript(code, id, id.slice(-2));
+				const preprocessed = preprocessScript(code, id, id.slice(-2));
 
 				return {
 					code: preprocessed.code,
