+++ src/public.d.ts
@@ -1,6 +1,7 @@
+import "./ambient.js"
+export * from "./index.js"
+
 export interface BreferConfig {
 	include?: string[];
 	exclude?: string[];
 }
-
-export * from "./index.js";
