+++ tsconfig.json
@@ -13,5 +13,6 @@
 		"allowJs": true,
 		"checkJs": true,
 		"allowSyntheticDefaultImports": true
-	}
+	},
+	"include": ["src"]
 }
